Extract attr helpers
