+++ crates/browser/src/form_controls/mod.rs
@@ -1,9 +1,9 @@
+mod dom;
 mod index;
-mod input_type;
 mod seed;
 
+pub use dom::{InputControlType, input_control_type};
 pub use index::FormControlIndex;
-pub use input_type::{InputControlType, input_control_type};
 pub use seed::seed_input_state_from_dom;
 
 #[cfg(test)]
